Make libphutil compatible with utf8mb4
Summary:
Ref T1191. Two changes here:

  - Try to use the `utf8mb4` character set. Fall back to `utf8` if it's not available. This may be slightly worse for old MySQL, but is far, far easier than trying to keeep a cache around somewhere with availability information.
  - Prepare to transition to `utf8mb4`. I'm just going to switch this over in the future. I considered trying to save the migration state in the environment somewhere, but it gets really, really messy because we have to inspect the migration state in order to build a connection object. There's also no advantage in the overwhelming majority of cases. We can do something more messy here eventually if it ends up being more complicated than I think.

Test Plan: Wrote comments full of emoji locally.

Reviewers: btrahan

Reviewed By: btrahan

Subscribers: epriestley

Maniphest Tasks: T1191

Differential Revision: https://secure.phabricator.com/D10604

+++ src/aphront/storage/connection/mysql/AphrontBaseMySQLDatabaseConnection.php
@@ -20,7 +20,7 @@ abstract class AphrontBaseMySQLDatabaseConnection
   abstract protected function freeResult($result);
 
   public function __construct(array $configuration) {
-    $this->configuration  = $configuration;
+    $this->configuration = $configuration;
   }
 
   public function __clone() {
@@ -324,16 +324,32 @@ abstract class AphrontBaseMySQLDatabaseConnection
    * can lead to data loss and security problems.
    */
   protected function validateUTF8String($string) {
-    if (phutil_is_utf8_with_only_bmp_characters($string)) {
-      return;
-    }
+    // TODO: Make this `true` eventually, once we make storage adjustment
+    // mandatory. For now, you can set it to `true` to test things.
+    $has_utf8mb4 = false;
 
-    throw new AphrontCharacterSetQueryException(
-      pht(
-        'Attempting to construct a query containing characters outside of '.
-        'the Unicode Basic Multilingual Plane. MySQL will silently truncate '.
-        'this data if it is inserted into a `utf8` column. Use the `%%B` '.
-        'conversion to escape binary strings data.'));
+    if ($has_utf8mb4) {
+      if (phutil_is_utf8($string)) {
+        return;
+      }
+
+      throw new AphrontCharacterSetQueryException(
+        pht(
+          'Attempting to construct a query using a non-utf8 string when '.
+          'utf8 is expected. Use the `%%B` conversion to escape binary '.
+          'strings data.'));
+    } else {
+      if (phutil_is_utf8_with_only_bmp_characters($string)) {
+        return;
+      }
+
+      throw new AphrontCharacterSetQueryException(
+        pht(
+          'Attempting to construct a query containing characters outside of '.
+          'the Unicode Basic Multilingual Plane. MySQL will silently truncate '.
+          'this data if it is inserted into a `utf8` column. Use the `%%B` '.
+          'conversion to escape binary strings data.'));
+    }
   }
 
 }
